consertando back-end

+++ src/main/java/com/matemagicos/biblioteca/repository/JogoRepository.java
@@ -3,5 +3,5 @@ package com.matemagicos.biblioteca.repository;
 import com.matemagicos.biblioteca.models.Jogo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JogoRepository extends JpaRepository<Jogo, Long> {
+public interface JogoRepository extends JpaRepository<Jogo, Integer> {
 }
