+++ src/main/java/com/matemagicos/biblioteca/repository/PontuacaoRepository.java
@@ -3,5 +3,5 @@ package com.matemagicos.biblioteca.repository;
 import com.matemagicos.biblioteca.models.PontuacaoHistorico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PontuacaoRepository extends JpaRepository<PontuacaoHistorico, Long> {
+public interface PontuacaoRepository extends JpaRepository<PontuacaoHistorico, Integer> {
 }
